+++ land-queue/.gitignore
@@ -1,4 +1,3 @@
-logs/webhooks.json
 *.pem
 
 # Byte-compiled / optimized / DLL files
