--- land-queue/config.py
@@ -1,5 +1,0 @@
-import os
-
-WEBHOOKS_LOG_FILE = "logs/webhooks.json"
-KEY_FILE = "monorepoapptest.private-key.pem"
-APP_ID = os.environ.get("APP_ID", "64638")
