+++ land-queue/main.py
@@ -1,51 +1,10 @@
-import json
-from flask import Flask, request, redirect
+from flask import Flask
 
-from land_queue.handlers.pull_request_handler import PullRequestHandler
-from land_queue.handlers.installation_handler import InstallationHandler
-from land_queue.handlers.integration_handler import IntegrationHandler
-from land_queue.config import Config
-
+from land_queue.views import views as bp_views
 
 app = Flask(__name__)
+app.register_blueprint(bp_views)
 
 
-WEBHOOK_EVENT_HANDLERS = {
-    'installation': InstallationHandler(),
-    'integration_installation': InstallationHandler(),
-    'pull_request': PullRequestHandler(Config)
-}
-
-
-@app.route('/webhooks', methods=['POST'])
-def webhooks():
-    github_event = request.headers.get('X-GitHub-Event')
-
-    if github_event in WEBHOOK_EVENT_HANDLERS:
-        handler = WEBHOOK_EVENT_HANDLERS[github_event]
-        handler.process(request.data)
-    else:
-        print(f'Ups. Unhandleable Github event: {github_event}!')
-
-    return "OK"
-
-
-@app.route('/')
-def index():
-    return f'<tt><big><a href="/install">Install {Config.GH_APP_SLUG} to your repo.</a></big></tt>'
-
-
-@app.route('/install')
-def install():
-    install_url = f'https://github.com/apps/{Config.GH_APP_SLUG}/installations/new'
-    return redirect(install_url)
-
-
-# https://cloud.google.com/appengine/docs/standard/python3/configuring-warmup-requests
-@app.route('/_ah/warmup')
-def warmup():
-    return '', 200, {}
-
-
-if __name__ == '__main__':
-    app.run(host='127.0.0.1', debug=True)
+if __name__ == "__main__":
+    app.run(host="127.0.0.1", debug=True)
